+++ src/main/java/com/template/server/domain/image/service/ImageGenerationService.java
@@ -17,9 +17,10 @@ import com.template.server.domain.member.repository.MemberRepository;
 import com.template.server.global.config.OpenAiConfig;
 import com.template.server.global.error.exception.BusinessLogicException;
 import com.template.server.global.error.exception.ExceptionCode;
-import com.template.server.global.util.LanguageDiscriminationUtils;
-import com.template.server.global.util.RateLimiterManager;
-import com.template.server.global.util.Translate;
+import com.template.server.global.util.aws.service.S3UploadService;
+import com.template.server.global.util.translate.LanguageDiscriminationUtils;
+import com.template.server.global.util.ratelimit.RateLimiterManager;
+import com.template.server.global.util.translate.Translate;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
