+++ src/test/java/com/template/server/domain/image/service/ImageGenerationServiceTest.java
@@ -13,7 +13,8 @@ import com.template.server.domain.member.entity.Member;
 import com.template.server.domain.member.repository.MemberRepository;
 import com.template.server.global.error.exception.BusinessLogicException;
 import com.template.server.global.error.exception.ExceptionCode;
-import com.template.server.global.util.RateLimiterManager;
+import com.template.server.global.util.aws.service.S3UploadService;
+import com.template.server.global.util.ratelimit.RateLimiterManager;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
